requirements.txt
telethon>=1.34.0
requests>=2.31.0
aiohttp>=3.9.0
feedparser>=6.0.10
fastapi>=0.110.0
uvicorn>=0.29.0
tweepy>=4.14.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
humanize>=4.9.0
